Gem Putty Repair Material Change & AOEHarvest code base.
Gem Putty is now a global repair material, it uses 20 levels and adds
9110 durability to any repairable item in an anvil.
Additionally it can also be used to change the name of most any item
in an anvil for 5 levels.


AOEHarvest code added in for the test tool, not curently working as
I probably missed something basic, but no errors according to Eclipse at least!


Signed-off-by: ReignOfMagic <nelriana@gmail.com>